Extract enterprise audit redaction module
